Backend/secure_api/requirements.txt
alembic==1.18.4
fastapi==0.135.1
passlib==1.7.4
pydantic==2.12.5
pydantic_settings==2.13.1
python_jose==3.3.0
SQLAlchemy==2.0.37
starlette==0.52.1
